fixed feature GET
